Improve handling of nested lists
This is specifically to improve the formatting of the output when
a bullet point list is included in a text item list.  The bullet
point items should be at the indent level of the text item rather
than the outermost level.  (Other nested configurations were not
tested.)

To allow nesting, the "InOver" flag was removed and the "Indent"
level is now increment and decremented rather than being set to 0
and 3.  (The "3" was moved to the "_update_prologue" method so
that it does not need to be repeated -- the "Indent" level is
now changed in steps of one.)

The other thing preventing nesting was that Text::Wrap::wrap was
always called with no first line indent.  (The configured indent
be applied only to subsequent lines.)  This gave the desired formatting
for text item list terms, but is not suitable for use with nesting.
Instead the "Indent" level is temporarily reduced while processing the
text item itself.  At the end of the text item, "_update_prologue"
is called to write text at the temporary indent level, and then
"Indent" is restored.